# web/openapi.yaml
openapi: 3.1.0
info:
  title: Arc Agentic API
  version: "0.1.0"
  description: |
    Catalog of paid services on Arc Agentic — a stablecoin-native
    agent marketplace running on **Arc** (Circle's stablecoin Layer 1)
    via **x402** + **Circle Gateway**.

    ## Payment model
    Every endpoint is gated by HTTP 402. The first request returns a
    `402 Payment Required` with a `PAYMENT-REQUIRED` header describing
    the price and accepted networks. The agent signs a payment header
    (`PAYMENT-SIGNATURE`) and retries — the seller's middleware verifies
    against Circle Gateway and only then runs the handler.

    All prices are in **USDC** on **Arc Testnet** (CAIP-2 id `eip155:5042002`).
    Sub-cent amounts are settled in batched nanopayments — per-call gas
    is effectively zero.

    ## For AI agents
    Drop the URL pattern of any endpoint into your prompt; the schema
    here mirrors what each endpoint returns. Combine multiple endpoints
    using the bundle recipes in `web/data.js` → `BUNDLES`.

    Reference orchestrator: `agent/orchestrator.js` (runs any bundle id).

  contact:
    name: Arc Agentic
    url: https://github.com/anhdopc/liftup-agents
  license:
    name: MIT
servers:
  - url: https://{subdomain}.arc-agentic.dev
    description: Per-service domains on Arc Testnet
    variables:
      subdomain:
        default: data
        enum: [search, infer, data, chain, risk, fin, news, defi]
...
components:
  responses:
    PaymentRequired:
      description: |
        Payment required. Sign the `accepts` payload with your Agent Wallet
        and retry with `PAYMENT-SIGNATURE` header.
      headers:
        PAYMENT-REQUIRED:
          schema:
            type: string
          description: JSON-encoded payment requirement (see schema).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PaymentRequiredBody"
    Success:
      description: Payment verified, handler ran, data returned.
      content:
        application/json:
          schema:
            type: object
            description: Endpoint-specific shape. See each path for details.
            additionalProperties: true
  schemas:
    PaymentRequiredBody:
      type: object
      required: [accepts, resource]
      properties:
        accepts:
          type: array
          items:
            type: object
            required: [scheme, price, network, payTo]
            properties:
              scheme:      { type: string, enum: [exact, permit2], example: "exact" }
              price:       { type: string, example: "$0.001" }
              network:     { type: string, example: "eip155:5042002" }
              payTo:       { type: string, example: "0xSellerWalletAddress" }
              description: { type: string }
              mimeType:    { type: string, example: "application/json" }
              expiresIn:   { type: integer, description: "Seconds until quote expires.", example: 300 }
        resource:
          type: string
          example: "https://data.arc-agentic.dev/v1/price/ETH"
  parameters:
    Token:
      name: token
      in: path
      required: true
      schema: { type: string }
      example: ETH
...
  securitySchemes:
    x402:
      type: http
      scheme: bearer
      bearerFormat: PAYMENT-SIGNATURE
      description: |
        x402 payment header. The first request to any path returns 402.
        Sign the `accepts` quote with your Agent Wallet, attach the
        signed payload as `PAYMENT-SIGNATURE`, and retry.
security:
  - x402: []

# ════════════════════════════════════════════════════════════════
#  PATHS
# ════════════════════════════════════════════════════════════════
paths:
...
  /v1/price/{token}:
    get:
      tags: [Data]
      summary: Real-time price for any token by symbol or address.
      description: Price $0.001.
      parameters: [{ $ref: "#/components/parameters/Token" }]
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "402": { $ref: "#/components/responses/PaymentRequired" }
...